+++ frontend/css/style.css
@@ -176,6 +176,7 @@ main {
   height: 60px;
   border-radius: 50px;
   padding-right: 70px;
+  padding-left: 55px;
   border: 1px solid #e0e0e0;
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.08);
 }
@@ -203,6 +204,25 @@ main {
   transition: background-color 0.2s ease-in-out;
 }
 
+.health-query-btn-add {
+  position: absolute;
+  left: 28px;
+  padding: 3px;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 35px;
+  height: 35px;
+  border-radius: 50%;
+  background-color: #000000a6;
+  border: none;
+  color: black;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  box-shadow: 0 0 15px rgb(0 0 0 / 70%);
+  transition: background-color 0.2s ease-in-out;
+}
+
 .health-query-btn:hover {
   background-color: #d12134;
 }
@@ -211,6 +231,39 @@ main {
   font-size: 1.5rem;
   line-height: 1;
 }
+
+#dropzoneWrapper {
+  position: absolute;
+  z-index: 9999;
+  min-width: 250px;
+  max-width: 400px;
+  background: white;
+  border-radius: 8px;
+  box-shadow: 0 4px 12px rgba(0,0,0,0.15);
+  padding: 12px;
+  display: none;
+}
+
+.dz-progress {
+  top: 100%!important;
+}
+
+#imageDropzone {
+  padding: 0;
+}
+
+.dz-remove {
+      opacity: 1;
+    z-index: 21;
+    position: absolute;
+    left: 100%;
+    top: 0%;
+    margin-top: -8px;
+    overflow: hidden;
+    color: red;
+    font-size: 20px!important;
+}
+
 
 /* ---- CAIXA DE SUGESTÕES ---- */
 .suggestion-box {
@@ -252,9 +305,9 @@ main {
 /* ---- RODAPÉ ---- */
 .main-footer {
   margin-top: 20px;
-  padding: 1em;
+  padding: 0em;
   color: #000;
-  text-align: center;
+  text-align: start;
   position: fixed;
   bottom: 0px;
 }
@@ -320,7 +373,7 @@ body.chat-view-active #content {
 }
 
 #chat-container {
-  margin: 0!important;
+  margin: 0 !important;
   display: flex;
   width: 100%;
   flex-direction: column;
@@ -330,11 +383,11 @@ body.chat-view-active #content {
 }
 
 .push-bottom {
-  height: 0px!important;
-  position: fixed!important;
-  bottom: 0!important;
-  padding: 0!important;
-  border: 0!important;
+  height: 0px !important;
+  position: fixed !important;
+  bottom: 0 !important;
+  padding: 0 !important;
+  border: 0 !important;
   margin-left: -1.5em;
 }
 
@@ -434,16 +487,28 @@ body.chat-view-active #content {
   animation-delay: 0.9s;
 }
 
+.main-footer>p {
+  text-align: start;
+  margin-bottom: 0;
+}
+
 
 @media (max-width: 1024px) {
   #content {
     max-width: 100%;
     width: 100%;
     margin: 0px !important;
+    height: 95%;
+    padding-top: 0;
   }
+
+  #chat-container {
+    padding: 0 !important;
+  }
+
   .container {
-    margin-top: 0!important;
-    margin-bottom: 0!important;
+    margin-top: 0 !important;
+    margin-bottom: 0 !important;
   }
 
   div.w-50 {
@@ -456,7 +521,16 @@ body.chat-view-active #content {
 
   .health-query-container {
     width: 100% !important;
-    max-width: 90% !important;
+    max-width: 100% !important;
+    padding: 15px 0;
+  }
+
+  .health-query-btn-add {
+    left: 12px;
+  }
+
+  .health-query-btn {
+    right: 12px;
   }
 
   .welcome-title {
@@ -464,17 +538,24 @@ body.chat-view-active #content {
   }
 
   .main-footer>p {
-    margin-bottom: 3px;
+    text-align: start;
+    margin-bottom: 0;
+    margin-top: 1em !important;
   }
 
   .main-footer {
     padding: 10px 0 0 10px;
   }
 
+  .hide-mobile {
+    display: none;
+  }
+
   .aboutus-bg {
     padding-top: 1.5em !important;
     margin-top: 0px !important;
   }
+
   body.chat-view-active .health-query-container {
     transform: translateX(-105%);
   }
