+++ frontend/index.html
@@ -23,22 +23,28 @@
     <div id="content"></div>
   </main>
 
-  <footer class="main-footer">
+  <footer class="main-footer pb-0">
     <p>Made with care by Baymax <i class="bi bi-stars" style="color: #e62e41;"></i></p>
+    <p class="small mb-0 hide-mobile text-muted">Do not rely on Baymax to make decisions regarding your health and medical care. All answers are AI generated and prone to erros. For research purposes only.</p>
   </footer>
-  
+
   <script type="module" src="js/index.js" defer></script>
 
   <script
     src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
     defer
   ></script>
+
+  <script src="https://unpkg.com/dropzone@5/dist/min/dropzone.min.js"></script>
+  <link rel="stylesheet" href="https://unpkg.com/dropzone@5/dist/min/dropzone.min.css" type="text/css" />
+  <script src="https://unpkg.com/@popperjs/core@2"></script>
   
   <script src="https://cdn.jsdelivr.net/npm/fullcalendar@6.1.15/index.global.min.js"></script>
 
   <script src="https://unpkg.com/leaflet/dist/leaflet.js" defer></script>
   
   <script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
+  
 
   <noscript>
     <div class="container my-4">
