+++ frontend/js/routes.js
@@ -1,4 +1,4 @@
-import renderHome from "./views/pages/home.js";
+import renderHome from "./views/pages/baymax.js";
 import renderLearnMore from "./views/pages/learnMore.js";
 import renderMedications from "./views/pages/medications.js";
 import renderAboutUs from "./views/pages/aboutUs.js";
