+++ frontend/js/services/handle-baymax-question.js
@@ -1,22 +1,26 @@
 const baymaxUrl = "http://localhost:8080/baymax/api/ask-baymax";
+const baymaxImageUrl = "http://localhost:8080/baymax/api/show-baymax"; // separate endpoint for image
 
-async function handleBaymaxQuestion(question) {
+async function handleBaymaxQuestion(question, imageBase64 = null) {
   try {
-    const response = await fetch(baymaxUrl, {
+    const url = imageBase64 ? baymaxImageUrl : baymaxUrl;
+
+    const payload = { question };
+    if (imageBase64) payload.image = imageBase64.split(',')[1];
+
+    const response = await fetch(url, {
       method: "POST",
       headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ question })
+      body: JSON.stringify(payload),
     });
 
     if (!response.ok) {
       throw new Error(`Server error: ${response.status} ${response.statusText}`);
     }
 
-    
     const baymaxActualResponse = await response.json();
     const content = baymaxActualResponse.output?.content || "";
 
-    //console.log("Baymax says:", content);
     return content;
   } catch (error) {
     console.error("Error talking to Baymax:", error);
@@ -24,4 +28,4 @@ async function handleBaymaxQuestion(question) {
   }
 }
 
-export {handleBaymaxQuestion};
+export { handleBaymaxQuestion };
