--- frontend/js/views/components/baymax-resp/handle-baymax-response.js
@@ -1,7 +1,0 @@
-
-
-
-function handleBaymaxResponse(response){
-
-
-}
