+++ frontend/js/views/components/commons/button.js
@@ -1,8 +1,10 @@
 
-export default function button(text, onClick, additionalClass = "") {
+function button(text, onClick, additionalClass = "") {
   const btn = document.createElement("button");
   btn.textContent = text;
   btn.className = ["btn", additionalClass].filter(Boolean).join(" ");
   btn.addEventListener("click", onClick);
   return btn;
 }
+
+export {button};
