+++ frontend/js/views/components/svg.js
@@ -1,0 +1,2 @@
+const clipSvg = `<svg width="64px" height="64px" viewBox="0 0 24 24" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" fill="#ffffff"></g><g id="SVGRepo_iconCarrier"> <title>attachment_line</title> <g id="页面-1" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd"> <g id="File" transform="translate(-48.000000, 0.000000)"> <g id="attachment_line" transform="translate(48.000000, 0.000000)"> <path d="M24,0 L24,24 L0,24 L0,0 L24,0 Z M12.5934901,23.257841 L12.5819402,23.2595131 L12.5108777,23.2950439 L12.4918791,23.2987469 L12.4918791,23.2987469 L12.4767152,23.2950439 L12.4056548,23.2595131 C12.3958229,23.2563662 12.3870493,23.2590235 12.3821421,23.2649074 L12.3780323,23.275831 L12.360941,23.7031097 L12.3658947,23.7234994 L12.3769048,23.7357139 L12.4804777,23.8096931 L12.4953491,23.8136134 L12.4953491,23.8136134 L12.5071152,23.8096931 L12.6106902,23.7357139 L12.6232938,23.7196733 L12.6232938,23.7196733 L12.6266527,23.7031097 L12.609561,23.275831 C12.6075724,23.2657013 12.6010112,23.2592993 12.5934901,23.257841 L12.5934901,23.257841 Z M12.8583906,23.1452862 L12.8445485,23.1473072 L12.6598443,23.2396597 L12.6498822,23.2499052 L12.6498822,23.2499052 L12.6471943,23.2611114 L12.6650943,23.6906389 L12.6699349,23.7034178 L12.6699349,23.7034178 L12.678386,23.7104931 L12.8793402,23.8032389 C12.8914285,23.8068999 12.9022333,23.8029875 12.9078286,23.7952264 L12.9118235,23.7811639 L12.8776777,23.1665331 C12.8752882,23.1545897 12.8674102,23.1470016 12.8583906,23.1452862 L12.8583906,23.1452862 Z M12.1430473,23.1473072 C12.1332178,23.1423925 12.1221763,23.1452606 12.1156365,23.1525954 L12.1099173,23.1665331 L12.0757714,23.7811639 C12.0751323,23.7926639 12.0828099,23.8018602 12.0926481,23.8045676 L12.108256,23.8032389 L12.3092106,23.7104931 L12.3186497,23.7024347 L12.3186497,23.7024347 L12.3225043,23.6906389 L12.340401,23.2611114 L12.337245,23.2485176 L12.337245,23.2485176 L12.3277531,23.2396597 L12.1430473,23.1473072 Z" id="MingCute" fill-rule="nonzero"> </path> <path d="M18.7103,17.5652 C20.37,15.9055 20.37,13.2145 18.7103,11.5548 L12.1696,5.01406 C11.7791,4.62353 11.7791,3.99037 12.1696,3.59984 C12.5601,3.20932 13.1933,3.20932 13.5838,3.59984 L20.1245,10.1406 C22.5653,12.5814 22.5653,16.5386 20.1245,18.9794 C17.6838,21.4202 13.7265,21.4202 11.2857,18.9794 L3.33178,11.0255 C1.57385,9.26757 1.57385,6.4174 3.33178,4.65947 C5.08971,2.90154 7.93988,2.90154 9.69781,4.65947 L17.6507,12.6123 C18.7252,13.6869 18.7252,15.429 17.6507,16.5035 C16.5762,17.578 14.834,17.578 13.7595,16.5035 L6.51272,9.2567 C6.1222,8.86617 6.1222,8.23301 6.51272,7.84248 C6.90325,7.45196 7.53641,7.45196 7.92694,7.84248 L15.1737,15.0893 C15.4672,15.3828 15.943,15.3828 16.2365,15.0893 C16.5299,14.7958 16.5299,14.32 16.2365,14.0266 L8.2836,6.07368 C7.30671,5.0968 5.72287,5.0968 4.74599,6.07368 C3.76911,7.05056 3.76911,8.6344 4.74599,9.61129 L12.6999,17.5652 C14.3596,19.2249 17.0506,19.2249 18.7103,17.5652 Z" fill="#d1d4d6ff"> </path> </g> </g> </g> </g></svg>`;
+export {clipSvg};
